# api/openapi.yaml
openapi: 3.1.0
info:
  title: clanker-proxy API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    One API with four access levels, declared per operation and checked
    before any handler runs. **owner** (the default): the daemon's owner,
    with the token in `<dir>/owner.token`; cpctl uses it. **agent**: a token
    the owner creates for an agent working for them; it reads and acts on
    existing threads and nothing else. **peer**: another daemon, with the
    secret it shares with this one; it identifies which peer is calling.
    **public**: no token; only `getHealth` and `requestPeering`.
    This spec is the source of truth for `api/rest` (server and client).
servers:
  - url: /
    description: The daemon
security:
  - ownerToken: []
...
paths:
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List webhooks
      description: >-
        List webhooks.
      responses:
        '200':
          description: Configured endpoints, without destination credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookList'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Create a webhook
      description: >-
        Only future matching events are queued; secrets are never returned.
      requestBody:
        required: true
        description: Settings and signing key for the new endpoint.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: The new webhook, without destination credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/webhooks/{name}:
    parameters:
      - $ref: '#/components/parameters/WebhookName'
    get:
      operationId: getWebhook
      tags: [webhooks]
      summary: Read a webhook
      description: >-
        Read a webhook.
      responses:
        '200':
          description: The endpoint settings, without destination credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: updateWebhook
      tags: [webhooks]
      summary: Update a webhook
      description: >-
        Changes only supplied settings. The destination type is immutable. Omit
        url, headers or secret to keep them; an empty secret disables signing and
        empty headers remove custom headers. Pending deliveries use the current
        destination and credentials. Disabled webhooks queue no new events and
        pause existing deliveries; an in-flight request may finish. Any update
        clears the endpoint's failure backoff (pausedUntil).
      requestBody:
        required: true
        description: Settings to change; omitted fields are retained.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '200':
          description: The updated webhook, without destination credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a webhook
      description: >-
        Deletes configuration and delivery history, including pending deliveries. An in-flight request may finish.
      responses:
        '204':
          description: The webhook and its delivery history are gone.
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/agent-tokens:
    get:
      operationId: listAgentTokens
      tags: [tokens]
      summary: List agent tokens
      description: Every agent token, by name, without the tokens themselves.
      responses:
        '200':
          description: Agent tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTokenList'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createAgentToken
      tags: [tokens]
      summary: Create an agent token
      description: >-
        The token is returned once and only its hash is stored. It may list,
        read and act on threads, follow the stream and read the owner's name;
        every other operation refuses it with 403. Fails with 409 when the
        name is taken, and with 400 when `expiresAt` is not in the future.
      requestBody:
        description: The token's name and optional expiry.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTokenInput'
      responses:
        '201':
          description: The new token, shown only here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewAgentToken'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/agent-tokens/{name}:
    parameters:
      - name: name
        in: path
        required: true
        description: The agent token's name.
        schema:
          $ref: '#/components/schemas/Name'
    delete:
      operationId: deleteAgentToken
      tags: [tokens]
      summary: Revoke an agent token
      description: The token stops working at once.
      responses:
        '204':
          description: Revoked.
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/peers:
    get:
      operationId: listPeers
      tags: [peering]
      summary: List peers
      description: >-
        Every peer, requested or active.
      responses:
        '200':
          description: Every peer, by name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerList'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: addPeer
      tags: [peering]
      summary: Ask another daemon to become peers
      description: >-
        Makes a secret, sends it with a peering request to the daemon at
        `url`, and saves the peer as `requested`. Events to them queue until
        their owner approves; then they deliver. Fails with 502 when their
        daemon cannot be reached, and saves nothing. Fails with 409 when
        the name or URL already belongs to a peer; each daemon has one name
        here. Trailing slashes on daemon URLs are removed. The URL must reach
        the daemon directly; redirects are refused.
      requestBody:
        description: Who to ask, and where.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeerInput'
      responses:
        '201':
          description: The requested peer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Peer'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/peers/{name}:
    parameters:
      - $ref: '#/components/parameters/PeerName'
    get:
      operationId: getPeer
      tags: [peering]
      summary: Read a peer
      description: >-
        One peer, with the short code of the secret you share.
      responses:
        '200':
          description: The peer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Peer'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: removePeer
      tags: [peering]
      summary: Remove a peer
      description: >-
        Their secret stops working and undelivered events to them fail.
        Threads are kept, so their name can later go only to a peer at the
        same URL.
      responses:
        '204':
          description: The peer is gone.
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/stream:
    get:
      operationId: streamEvents
      tags: [threads]
      security:
        - ownerToken: []
        - agentToken: []
      summary: Stream thread changes
      description: >-
        Server-sent events. `thread`: a `ThreadSummary`, sent whenever a thread
        changes: an event sent or received, or a delivery that succeeded or
        failed. A `: ping` comment keeps idle connections open. Nothing is
        replayed on connect; list threads first, then follow the stream.
      responses:
        '200':
          description: Server-sent `thread` events.
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ThreadSummary'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    webhookSignature:
      type: apiKey
      in: header
      name: Webhook-Signature
      description: HMAC-SHA256 in Standard Webhooks v1 format, verified by the external receiver.
    ownerToken:
      type: http
      scheme: bearer
      description: The owner token, from `<dir>/owner.token`.
    peerSecret:
      type: http
      scheme: bearer
      description: The secret shared with one peer.
    agentToken:
      type: http
      scheme: bearer
      description: An agent token the owner created; limited to existing threads.
  parameters:
    WebhookName:
      name: name
      in: path
      required: true
      description: The webhook name.
      schema:
        $ref: '#/components/schemas/Name'
...
    PeerName:
      name: name
      in: path
      required: true
      description: The peer's name.
      schema:
        $ref: '#/components/schemas/Name'
...
  responses:
    Problem:
      description: An error, as an RFC 9457 problem.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    WebhookEvents:
      type: array
      description: A sole '*' subscribes to all current and future event types; otherwise an explicit nonempty list.
      minItems: 1
      maxItems: 10
      uniqueItems: true
      items:
        type: string
        enum: ['*', thread.open, thread.reply, thread.ack, thread.needs-input, thread.resolve, thread.decline, thread.close, thread.reopen, thread.withdraw, peering.requested]
    WebhookOrigin:
      type: string
      description: Incoming events from peers, outgoing owner actions, or both. Peering requests are incoming.
      enum: [incoming, outgoing, both]
...
    WebhookURL:
      type: string
      description: >-
        A webhook destination: https, or http on a literal loopback address,
        without userinfo or fragment. Paths and queries may contain provider
        credentials, so the entire URL is write-only. Any host the daemon can
        reach is allowed. At most 2048 bytes. Validation errors never echo the URL.
      minLength: 1
      maxLength: 2048
      writeOnly: true
      examples: ['https://runner.example.com/hooks']
    WebhookType:
      type: string
      description: Destination format. Immutable after creation. Teams uses Workflows Adaptive Cards; apprise sends to a configured Apprise API endpoint.
      enum: [generic, discord, slack, teams, google-chat, mattermost, rocketchat, ntfy, gotify, apprise]
    WebhookHeaderName:
      type: string
      description: An HTTP header name. Transport, Content-Type, User-Agent, Idempotency-Key, Proxy-* and Webhook-* headers are reserved.
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9!#$%&''*+.^_`|~-]+$'
    WebhookHeaders:
      type: array
      description: Custom headers, including authentication. Names must be unique ignoring case. Values are write-only; empty replaces all headers with none.
      maxItems: 16
      items:
        type: object
        required: [name, value]
        properties:
          name:
            $ref: '#/components/schemas/WebhookHeaderName'
          value:
            type: string
            minLength: 1
            maxLength: 4096
            writeOnly: true
            description: A single-line value of at most 4096 bytes, without control characters.
    WebhookSecret:
      type: string
      description: Optional base64 encoding of a 32-byte signing key for generic destinations only. Empty disables signing. Write-only; never returned or logged.
      maxLength: 44
      pattern: '^([A-Za-z0-9+/]{43}=)?$'
      writeOnly: true
    Webhook:
      type: object
      description: >-
        Owner-configured notifications, without destination paths, queries,
        custom header values or signing keys.
      required: [name, type, destination, events, origin, enabled, signing, headerNames]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        type:
          $ref: '#/components/schemas/WebhookType'
        destination:
          type: string
          maxLength: 2048
          description: Destination scheme and host only; path and query are never returned.
        signing:
          type: boolean
          description: Whether generic notifications have a configured signing key.
        headerNames:
          type: array
          maxItems: 16
          items:
            $ref: '#/components/schemas/WebhookHeaderName'
        events:
          $ref: '#/components/schemas/WebhookEvents'
        origin:
          $ref: '#/components/schemas/WebhookOrigin'
        enabled:
          type: boolean
        pausedUntil:
          type: string
          format: date-time
          readOnly: true
          description: >-
            Set while the endpoint is backing off after failed attempts or a
            receiver's Retry-After; no delivery is attempted before it. Any
            update or manual retry clears it.
    WebhookCreate:
      type: object
      description: >-
        A new webhook. Generic signing is opt-in. Destination URL, custom header
        values and signing key are write-only. Chat destinations authenticate
        with their provider URL or custom headers, without a signing key.
      required: [name, type, url, events, origin, enabled]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        type:
          $ref: '#/components/schemas/WebhookType'
        url:
          $ref: '#/components/schemas/WebhookURL'
        events:
          $ref: '#/components/schemas/WebhookEvents'
        origin:
          $ref: '#/components/schemas/WebhookOrigin'
        enabled:
          type: boolean
        secret:
          $ref: '#/components/schemas/WebhookSecret'
        headers:
          $ref: '#/components/schemas/WebhookHeaders'
    WebhookUpdate:
      type: object
      additionalProperties: false
      description: >-
        Changes to a destination. Omitted settings are retained. Empty secret
        disables signing; empty headers remove custom headers. Type is immutable.
      properties:
        url:
          $ref: '#/components/schemas/WebhookURL'
        events:
          $ref: '#/components/schemas/WebhookEvents'
        origin:
          $ref: '#/components/schemas/WebhookOrigin'
        enabled:
          type: boolean
        secret:
          $ref: '#/components/schemas/WebhookSecret'
        headers:
          $ref: '#/components/schemas/WebhookHeaders'
    WebhookList:
      type: object
      description: Configured webhooks, sorted by name.
      required: [webhooks]
      properties:
        webhooks:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
...
    Name:
      type: string
      description: >-
        A participant's name: lower-case letters,
        digits and single inner hyphens, at most 39 characters.
      minLength: 1
      maxLength: 39
      pattern: '^[a-z0-9](-?[a-z0-9])*$'
      examples: [savid]
    Label:
      type: string
      description: A tag on a thread, such as a project or repository.
      minLength: 1
      maxLength: 40
      pattern: '^[a-z0-9][a-z0-9._/-]*$'
      examples: [infra]
    Labels:
      type: array
      description: A thread's labels, unique and sorted.
      maxItems: 10
      uniqueItems: true
      items:
        $ref: '#/components/schemas/Label'
    ID:
      type: string
      description: A lower-case UUID.
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      maxLength: 36
    Title:
      type: string
      description: A thread's one-line title, at most 200 characters.
      minLength: 1
      maxLength: 200
...
    Note:
      type: string
      description: A message for the other owner.
      maxLength: 500
    Code:
      type: string
      description: >-
        A shared secret's short code: the first 8 hex digits of its SHA-256.
        Both owners see the same code; comparing them over chat confirms a
        peering request came from who it says.
      pattern: '^[0-9a-f]{4}-[0-9a-f]{4}$'
      maxLength: 9
      examples: ['4f2a-91c3']
...
    Count:
      type: integer
      description: A count.
      format: int32
      minimum: 0
      maximum: 1000000
    DaemonURL:
      type: string
      format: uri
      description: >-
        A daemon's base URL: absolute http or https with a host, at most 512
        characters (checked by the server; ogen cannot bound a URI).
      maxLength: 512
      examples: ['https://cp.example.com']
...
    PeerInput:
      description: >-
        Who to ask to become peers.
      type: object
      required: [name, url]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        url:
          $ref: '#/components/schemas/DaemonURL'
        note:
          $ref: '#/components/schemas/Note'
    Peer:
      description: >-
        A daemon this one exchanges threads with.
      type: object
      required: [name, url, status, code, addedAt]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        url:
          $ref: '#/components/schemas/DaemonURL'
        status:
          type: string
          description: >-
            `requested`: waiting for their owner to approve. `active`: both
            sides accept the secret.
          enum: [requested, active]
        code:
          $ref: '#/components/schemas/Code'
        addedAt:
          type: string
          format: date-time
    AgentTokenInput:
      description: A new agent token.
      type: object
      required: [name]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        expiresAt:
          type: string
          format: date-time
          description: When it stops working. Omitted, it works until revoked.
    AgentTokenSummary:
      description: An agent token, without the token.
      type: object
      required: [name, createdAt]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        usedAt:
          type: string
          format: date-time
          description: Its last use, to the minute.
    NewAgentToken:
      description: A new agent token and the token itself, which is never shown again.
      type: object
      required: [name, token, createdAt]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        token:
          type: string
          description: The bearer token. Give it to the agent as CP_TOKEN.
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    AgentTokenList:
      description: Agent tokens, by name.
      type: object
      required: [tokens]
      properties:
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/AgentTokenSummary'
    PeerList:
      description: >-
        Peers, by name.
      type: object
      required: [peers]
      properties:
        peers:
          type: array
          items:
            $ref: '#/components/schemas/Peer'
    ThreadState:
      description: >-
        Where a thread is in its workflow. `closed`, `declined` and `withdrawn` end it; a closed thread can be reopened.
      type: string
      enum: [open, acked, needs-input, resolved, closed, declined, withdrawn]
    ThreadKind:
      description: >-
        `request`: a resolution is expected. `fyi`: none is; acknowledging it closes it.
      type: string
      enum: [request, fyi]
    Action:
      description: >-
        What an event does. The recipient acks, asks (`needs-input`), resolves and declines; the sender closes, reopens and withdraws; either comments.
      type: string
      enum: [open, comment, ack, needs-input, resolve, close, reopen, decline, withdraw]
...
    ThreadSummary:
      description: >-
        A thread without its events.
      type: object
      required: [id, title, kind, labels, sender, recipient, peer, role, state, myTurn, actions, openedAt, updatedAt, events, undelivered, failed]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        title:
          $ref: '#/components/schemas/Title'
        kind:
          $ref: '#/components/schemas/ThreadKind'
        labels:
          $ref: '#/components/schemas/Labels'
        sender:
          $ref: '#/components/schemas/Name'
        recipient:
          $ref: '#/components/schemas/Name'
        peer:
          $ref: '#/components/schemas/Name'
        role:
          type: string
          description: >-
            The owner's part. `sender`: they opened it and close, reopen or
            withdraw it. `recipient`: they were asked and ack, ask, resolve
            or decline it.
          enum: [sender, recipient]
        state:
          $ref: '#/components/schemas/ThreadState'
        turn:
          $ref: '#/components/schemas/Name'
        myTurn:
          type: boolean
          description: The thread is waiting on the owner.
        actions:
          type: array
          description: >-
            What the owner may do now, for actOnThread. `comment` is always
            among them.
          items:
            $ref: '#/components/schemas/Action'
        openedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        events:
          $ref: '#/components/schemas/Count'
        undelivered:
          $ref: '#/components/schemas/Count'
        failed:
          $ref: '#/components/schemas/Count'
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference